Retrieve DOJ Press Releases

This function retrieves press releases from the U.S. Department of Justice API. You can specify the number of results, search direction, and optional keyword filtering.

Usage

doj_press_releases(
  n_results = 50,
  search_direction = "DESC",
  keyword = NULL,
  clean = TRUE
)

Arguments

n_results

Integer. Number of press releases to retrieve. Default is 50.

search_direction

Character. "ASC" or "DESC" for ascending or descending sort by date. Default is "DESC".

keyword

Optional character string to filter results by keyword.

clean

Logical. If TRUE, cleans up the text fields. Default is TRUE.

Value

A cleaned dataframe containing DOJ press release metadata and content.

Examples

# \dontrun{
# Retrieve a single DOJ press release
result <- doj_press_releases(n_results = 1)
#> Warning: There were 2 warnings in `mutate()`.
#> The first warning was:
#> ℹ In argument: `created = anydate(as.numeric(created))`.
#> Caused by warning in `anydate()`:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
print(result)
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           body
#> 1 A 38-count superseding indictment (S4 indictment) was unsealed today charging 27 members of the Anti-Tren faction of the designated foreign terrorist organization Tren de Aragua (TDA or TdA) with offenses including racketeering conspiracy, murder in aid of racketeering in connection with an April 15, 2024, double murder in the Bronx, New York, murder-for-hire conspiracy, kidnapping in aid of racketeering, sex trafficking and sex trafficking conspiracy, and other offenses. Twenty-one of the 27 defendants charged in the S4 indictment were previously charged in a 12-count indictment. Of the six newly added defendants, five are now in federal custody.To date, the U.S. Attorney’s Office for the Southern District of New York has charged approximately 38 members or associates of TDA and Anti-Tren.“As alleged, these members of Anti-Tren, a splinter faction of the terrorist organization Tren de Aragua, planned and carried out a series of horrific crimes, including gunpoint robberies, murders, and the exploitation of vulnerable young women through sex trafficking,” said U.S. Attorney Jay Clayton for the Southern District of New York (SDNY).&nbsp;“Tren de Aragua is in the business of murder, sex trafficking, and&nbsp;intimidation, and they brought that business to New York while being unlawfully present in the United States. Today’s charges reflect our commitment to bankrupt TDA and bring its members to justice. The gang members charged today exerted ruthless control over sex trafficking victims through intimidation, brutality, and threats of violence against them and their loved ones — leaving lasting trauma in their wake. The women and men of the SDNY are&nbsp;unwavering in their commitment to bankrupt gangs that corrupt our neighborhoods, prey on the vulnerable, and pursue violence as a way of life.”“Homeland Security Investigations New York continues to stand at the forefront of investigations against vicious criminal enterprises like Tren de Aragua and Anti-Tren,” said Acting Executive Associate Director John A. Condon for Homeland Security Investigations (HSI). “Through their use of ruthless tactics and blatant disregard for human life, TdA and its offshoots are among the fastest emerging transnational criminal organizations to encroach upon American soil. HSI New York, through the Homeland Security Task Force, continues to target these gangs that seek to perpetrate destruction and terror in our communities. Together, alongside our law enforcement partners, we are committed to ensuring that no corners of the TdA and Anti-Tren enterprises are beyond the reach of justice.”“TdA and its faction Anti-Tren grew from a prison gang to a transnational criminal organization to a foreign terrorist organization. The violence and horrific acts as charged here including the double murder, murder for hire, kidnapping, and sex trafficking will be met with the full weight of our justice system,” said Co-Director Christopher Eason of Joint Task Force Vulcan (JTFV). “The charges against&nbsp;these TdA associates and Anti-Tren members are directly in line with JTFV’s mission: a collaborative, whole of government effort to destroy TdA and its factions. We are grateful for our partnership with the U.S. Attorney’s Office for the Southern District of New York, and our law enforcement partners who worked tirelessly to investigate and bring these important charges.”According to the allegations contained in the S4 indictment, Anti-Tren is a criminal organization almost exclusively comprised of former members and associates of TDA. Anti-Tren operated throughout New York City, including the boroughs of the Bronx and Queens, and in New Jersey, and elsewhere around the country, including in Illinois and Washington. The purposes of Anti-Tren included:Preserving and protecting the power and territory of Anti-Tren and its members and associates through acts involving murder, assault, other acts of violence, and threats of violence, including acts of violence and threats of violence directed at members and associates of TDA;Enriching the members and associates of Anti-Tren through, among other things:The unlawful smuggling of individuals, including women and girls from Venezuela, into the United States,The sex trafficking of young women or “multadas” who had been unlawfully smuggled into the United States,The trafficking of controlled substances, including “tusi,” andArmed robberies;Keeping victims and potential victims in fear of Anti-Tren and its members and associates through threats and acts of violence;Promoting and enhancing Anti-Tren and the reputation and activities of its members and associates;Providing assistance to members and associates of Anti-Tren who committed crimes for and on behalf of Anti-Tren, such as lodging and interstate transportation for members and associates of Anti-Tren to flee prosecution, or bail money for members or associates of Anti-Tren who are detained; andProtecting Anti-Tren and its members and associates from detection and prosecution by law enforcement authorities through acts of intimidation, threats, and violence against potential witnesses to crimes committed by members of Anti-Tren.Anti-Tren engaged in human smuggling and sex trafficking of young women or “multadas,” into the United States in exchange for debts that the young women would pay back by engaging in commercial sex work. Like TDA, members of Anti-Tren enforced compliance among these young women by, among other things:Threatening to kill the young women and their families,Assaulting the young women,Shooting or killing the young women, andTracking down and kidnapping the young women who tried to flee.Members of Anti-Tren also committed and conspired, attempted, and threatened to commit, acts of violence, including acts involving murder and assault, to protect and to expand Anti-Tren’s criminal operations, resolve disputes within Anti-Tren, to retaliate against rival organizations, including Tren de Aragua itself, and to maintain control over sex trafficking victims. Anti-Tren members and associates also trafficked controlled substances, committed robberies, and obtained, possessed, trafficked, and used firearms and ammunition.The new charges in the S4 indictment allege, among other things, that:On or about April 15, 2024, Yender Mata, Ervin Hernandez, also known as “Coco” and “Coquito,” and Kerlyn Nataliy Perez-Lopez, also known as “Mou,” conspired to murder and caused the murders of Jhombeyker Jose Bisbal Pina and Adrian Mendoza Isturiz, who were shot to death in the Bronx.In or about April 2025, Guillermo Enrique Freites-Velazquez and Roiman Noe Bello Ferrer conspired to kill an individual in New York (Intended Victim-1).From in or about February 2025 until in or about March 2025, David Valencia-De La Rosa and Johan Carlos Mujica-Urpin, also known as “Sobrino,” conspired to kill another individual in New York (Intended Victim-2).In or about April 2025, Roiman Noe Bello Ferrer and others conspired to kill an individual in Florida (Intended Victim-3).On or about Oct. 5, 2024, Jesus David Barrios Garcia, also known as “Morocho,” ordered Kellen Jaspe Bustamanate to shoot an Anti-Tren member in the leg as a form of gang punishment. On or about the same day, Barrios Garcia also ordered Enrique Gustavo Boada Yanez, also known as “Chino,” to shoot a different Anti-Tren member in the leg, also as gang punishment.In or about January 2025, Keiswuel Orlando Palacios-Milano, also known as “Oturache” and “Inmortalidad,” Anderson Smith Zambrano-Pacheco, and Yeferson Prieto Galviz, also known as “Flacot” and “Flacote,” conspired to commit and did commit a gunpoint home invasion robbery of an individual (Robbery Victim-1) and Robbery Victim-1’s family in Yonkers.In or about January 2025, Keiswuel Orlando Palacios-Milano, also known as “Oturache” and “Inmortalidad,” Anderson Zambrano-Pacheco, Mario Pereda, also known as “Cara de Hombre,” Wilfredo Jose Avendaño Carrizalez, also known as “Kabuvy,” and Carlos Gabriel Santos Mogollon, conspired to commit a gunpoint robbery of a drug trafficker (Robbery Victim-2) in the Bronx.In or about 2024, Luis Jose Velasquez-Hurtado, also known as “Chito,” engaged in sex trafficking of a young woman from Venezuela (Trafficking Victim-1) whom Velasquez-Hurtado threatened with a gun, caused to be kidnapped, and whose immigration documents Velasquez-Hurtado caused to be seized.In or about 2024, Jesus David Barrios Garcia, also known as “Morocho,” Enrique Gustavo Boada Yanez, also known as “Chino,” Jesus Ruben Lopez Gonzalez, also known as “Lopez,” and Danger Leoner Sanchez Alfonzo, also known as “Danger,” engaged in sex trafficking of another young woman from Venezuela (Trafficking Victim-2).A chart containing the names, charges, and maximum penalties for the defendants is set forth below.The maximum potential sentences in this case are prescribed by Congress and are provided here for informational purposes only, as any sentencing of the defendants will be determined by the judge.U.S. Attorney Clayton praised the outstanding investigative work of HSI New York, Seattle, Chicago, and Portland, and the NYPD. He also thanked Joint Task Force Vulcan; the Arapahoe County District Attorney’s Office; the Aurora Police Department in Aurora, Colorado; the New York/New Jersey Regional Fugitive Task Force of the U.S. Marshals Service (USMS); the U.S. Customs and Border Protection’s National Gang Unit and New York Human Intelligence Division; U.S. Immigration and Customs Enforcement’s New York and Chicago Enforcement and Removal Operations; the Bureau of Alcohol, Tobacco, Firearms, and Explosives (ATF); U.S. Marshals Service-District of Oregon; U.S. Customs and Border Protection-New York; the New York City Crime Analysis Center at the New York/New Jersey High Intensity Drug Trafficking Area; the Computer Crimes Unit at the NYPD; the King County, Washington Sheriff’s Office; and the Seattle Police Department.This case is a part of JTFV, which was created in 2019 to eradicate MS-13 and is now expanded to target Tren de Aragua. The JTFV is comprised of U.S. Attorney’s Offices across the country, including the Southern and Eastern Districts of New York; Eastern and Western Districts of North Carolina; Eastern and Western Districts of Virginia; Southern District of Florida; Eastern District of Texas; Western District of Oklahoma; Northern District of Indiana; and the District of Nevada, as well as the Executive Office for U.S. Attorneys, Department of Justice’s National Security Division and the Criminal Division. Additionally, the FBI, DEA, HSI, ATF, USMS, and the Federal Bureau of Prisons are essential law enforcement partners with JTFV.This case is part of Operation Take Back America, a nationwide initiative that marshals the full resources of the Justice Department to repel the invasion of illegal immigration, achieve the total elimination of cartels and transnational criminal organizations, and protect our communities from the perpetrators of violent crime. Operation Take Back America streamlines efforts and resources from the Department’s Organized Crime Drug Enforcement Task Forces and Project Safe Neighborhood.This case is being handled by JTFV and Assistant U.S. Attorneys Jun Xiang, Kathryn Wheelock, Timothy Ly, and Andrew K. Chan for the Southern District of New York are in charge of the prosecution.The charges contained in the indictment are merely accusations. All defendants are presumed innocent unless and until proven guilty beyond a reasonable doubt in a court of law.
#>   changed created       date number
#> 1    <NA>    <NA> 2026-02-13 26-149
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           teaser
#> 1 A 38-count superseding indictment (S4 indictment) was unsealed today charging 27 members of the Anti-Tren faction of the designated foreign terrorist organization Tren de Aragua (TDA or TdA) with offenses including racketeering conspiracy, murder in aid of racketeering in connection with an April 15, 2024, double murder in the Bronx, New York, murder-for-hire conspiracy, kidnapping in aid of racketeering, sex trafficking and sex trafficking conspiracy, and other offenses.&nbsp;Twenty-one of the 27 defendants charged in the S4 indictment were previously charged in a 12-count indictment.&nbsp;Of the six newly added defendants, five&nbsp;are now in federal custody.
#>                                                                                                                                                                                            title
#> 1 Twenty-Seven Members and Associates of Tren de Aragua Splinter Faction Anti-Tren Charged With 2024 Double Murder in New York and Sex Trafficking, Kidnapping, Racketeering, and Other Offenses
#>                                                                                                                         url
#> 1 https://www.justice.gov/opa/pr/twenty-seven-members-and-associates-tren-de-aragua-splinter-faction-anti-tren-charged-2024
#>                                   uuid
#> 1 a9a40f07-ddb2-4091-abd7-9d50f74f97c2
#>                                                                                                                                                 name
#> 1 Bureau of Alcohol, Tobacco, Firearms and Explosives (ATF), Office of the Deputy Attorney General, USAO - New York, Southern, U.S. Marshals Service
# }