Fetch U.S. DOJ Speeches

For accessing U.S. Department of Justice (DOJ) open data speeches.

Usage

doj_speeches(
    n_results = 50,
    search_direction = "DESC",
    keyword = NULL,
    clean = TRUE
  )

Arguments

n_results

The number of results to return. By default, 50 results are returned.

search_direction

Set to ASC or DESC to fetch data starting from the most or least recent entries. By default, data is sorted in descending order (the most recent entries).

keyword

Only return results containing a keyword. Keyword can be a single word, a list or words, or a regex separated by the pipe delimiter

clean

Set to FALSE to return data without additional cleaning. By default HTML tags, white space, line breaks, and other messy characters are removed. Epoch time is converted to year-month-day format.

Value

Dataframe containing API pull of U.S. DOJ speeches.

Examples

  # Return a dataframe with 50 of the latest records. Data is cleaned.
  latest_records <- doj_speeches()
  head(latest_records)
#>                                                                                       body
#> 1 Plain text version of this speech is not available, see the text accessible PDF version.
#> 2 Plain text version of this speech is not available, see the text accessible PDF version.
#> 3 Plain text version of this speech is not available, see the text accessible PDF version.
#> 4 Plain text version of this speech is not available, see the text accessible PDF version.
#> 5 Plain text version of this speech is not available, see the text accessible PDF version.
#> 6 Plain text version of this speech is not available, see the text accessible PDF version.
#>      changed    created           date teaser
#> 1 2020-01-22 2020-01-22 -2708440-05-05   <NA>
#> 2 2020-01-22 2020-01-22 -2616184-11-09   <NA>
#> 3 2020-01-15 2020-01-15 -2434736-07-10   <NA>
#> 4 2020-01-15 2020-01-15 -1732167-05-14   <NA>
#> 5 2020-01-23 2020-01-23 -1387033-05-09   <NA>
#> 6 2020-01-23 2020-01-23 -1288153-05-31   <NA>
#>                                                                      title
#> 1                               The Antitrust Laws, Their Past and Future 
#> 2 The Antitrust Laws as the First Line of Defense Against War Profiteering
#> 3                             The Antitrust Laws in the Present Emergency 
#> 4                                           Bigness and the Antitrust Laws
#> 5                              Strengthening Our Antitrust Administration 
#> 6                                                        Antitrust in 1954
#>                                                                                             url
#> 1                       https://www.justice.gov/atr/speech/antitrust-laws-their-past-and-future
#> 2 https://www.justice.gov/atr/speech/antitrust-laws-first-line-defense-against-war-profiteering
#> 3                           https://www.justice.gov/atr/speech/antitrust-laws-present-emergency
#> 4                                 https://www.justice.gov/atr/speech/bigness-and-antitrust-laws
#> 5                 https://www.justice.gov/atr/speech/strengthening-our-antitrust-administration
#> 6                                             https://www.justice.gov/atr/speech/antitrust-1954
#>                                   uuid                                vuuid
#> 1 ae2184cc-8734-4a59-9143-0a2d9016ee64 4bdbb626-d275-4b8d-98ea-6c4e5a29f65d
#> 2 819b8136-af99-4ce9-8331-6961b249ab75 3f1e8a12-b883-434a-b1d6-c7d7e17725bb
#> 3 67800d7a-4017-4a0b-8bd1-80b3ecf16c44 7f627aa8-a64f-46e4-b158-91cc3aecff3e
#> 4 e95aead0-79bd-462b-9f2c-75803b156bb4 fd96fd3f-6615-480d-a20d-d9c916063f09
#> 5 2df1540c-2863-45bd-b2c8-9bee042041e0 b43ee9ec-f23d-4cea-bd02-d32cf21c9923
#> 6 3d70829a-d269-46c2-b6b8-3f27e1814fc9 98d8a415-0f50-44b2-ba5f-64cbd83658dd
#>   location.country location.administrative_area location.locality
#> 1               US                           DC        Washington
#> 2               US                           NJ     Atlantic City
#> 3               US                           CA     Yosemite Park
#> 4               US                           NY          New York
#> 5               US                           IL           Chicago
#> 6               US                           NY          New York
#>   location.postal_code location.thoroughfare location.sub_premise
#> 1                 <NA>                  <NA>                   NA
#> 2                 <NA>                  <NA>                   NA
#> 3                 <NA>                  <NA>                   NA
#> 4                 <NA>                  <NA>                   NA
#> 5                 <NA>                  <NA>                   NA
#> 6                 <NA>                  <NA>                   NA
#>   location.phone_number location.phone_number_extension location.mobile_number
#> 1                  <NA>                            <NA>                   <NA>
#> 2                  <NA>                            <NA>                   <NA>
#> 3                  <NA>                            <NA>                   <NA>
#> 4                  <NA>                            <NA>                   <NA>
#> 5                  <NA>                            <NA>                   <NA>
#> 6                  <NA>                            <NA>                   <NA>
#>   location.fax_number               name
#> 1                <NA> Antitrust Division
#> 2                <NA> Antitrust Division
#> 3                <NA> Antitrust Division
#> 4                <NA> Antitrust Division
#> 5                <NA> Antitrust Division
#> 6                <NA> Antitrust Division

  # Return a dataframe with 15 of the earliest records. Data is not cleaned.
  earliest_records <- doj_speeches(n_results = 15, search_direction = "ASC", clean = FALSE)
  head(earliest_records)
#>                                                                                                                                                body
#> 1 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1237111/download">text accessible PDF version</a>.</p>\n
#> 2 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1237046/download">text accessible PDF version</a>.</p>\n
#> 3 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1234841/download">text accessible PDF version</a>.</p>\n
#> 4 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1234831/download">text accessible PDF version</a>.</p>\n
#> 5 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1237521/download">text accessible PDF version</a>.</p>\n
#> 6 <p>Plain text version of this speech is not available, see the <a href="/atr/speech/file/1237516/download">text accessible PDF version</a>.</p>\n
#>      changed    created           date teaser
#> 1 2020-01-22 2020-01-22 -2708440-05-05     NA
#> 2 2020-01-22 2020-01-22 -2616184-11-09     NA
#> 3 2020-01-15 2020-01-15 -2434736-07-10     NA
#> 4 2020-01-15 2020-01-15 -1732167-05-14     NA
#> 5 2020-01-23 2020-01-23 -1387033-05-09     NA
#> 6 2020-01-23 2020-01-23 -1288153-05-31     NA
#>                                                                      title
#> 1                               The Antitrust Laws, Their Past and Future 
#> 2 The Antitrust Laws as the First Line of Defense Against War Profiteering
#> 3                             The Antitrust Laws in the Present Emergency 
#> 4                                           Bigness and the Antitrust Laws
#> 5                              Strengthening Our Antitrust Administration 
#> 6                                                        Antitrust in 1954
#>                                                                                             url
#> 1                       https://www.justice.gov/atr/speech/antitrust-laws-their-past-and-future
#> 2 https://www.justice.gov/atr/speech/antitrust-laws-first-line-defense-against-war-profiteering
#> 3                           https://www.justice.gov/atr/speech/antitrust-laws-present-emergency
#> 4                                 https://www.justice.gov/atr/speech/bigness-and-antitrust-laws
#> 5                 https://www.justice.gov/atr/speech/strengthening-our-antitrust-administration
#> 6                                             https://www.justice.gov/atr/speech/antitrust-1954
#>                                   uuid                                vuuid
#> 1 ae2184cc-8734-4a59-9143-0a2d9016ee64 4bdbb626-d275-4b8d-98ea-6c4e5a29f65d
#> 2 819b8136-af99-4ce9-8331-6961b249ab75 3f1e8a12-b883-434a-b1d6-c7d7e17725bb
#> 3 67800d7a-4017-4a0b-8bd1-80b3ecf16c44 7f627aa8-a64f-46e4-b158-91cc3aecff3e
#> 4 e95aead0-79bd-462b-9f2c-75803b156bb4 fd96fd3f-6615-480d-a20d-d9c916063f09
#> 5 2df1540c-2863-45bd-b2c8-9bee042041e0 b43ee9ec-f23d-4cea-bd02-d32cf21c9923
#> 6 3d70829a-d269-46c2-b6b8-3f27e1814fc9 98d8a415-0f50-44b2-ba5f-64cbd83658dd
#>   location.country location.administrative_area location.locality
#> 1               US                           DC        Washington
#> 2               US                           NJ     Atlantic City
#> 3               US                           CA     Yosemite Park
#> 4               US                           NY          New York
#> 5               US                           IL           Chicago
#> 6               US                           NY          New York
#>   location.postal_code location.thoroughfare location.sub_premise
#> 1                 <NA>                  <NA>                   NA
#> 2                 <NA>                  <NA>                   NA
#> 3                 <NA>                  <NA>                   NA
#> 4                 <NA>                  <NA>                   NA
#> 5                 <NA>                  <NA>                   NA
#> 6                 <NA>                  <NA>                   NA
#>   location.phone_number location.phone_number_extension location.mobile_number
#> 1                  <NA>                            <NA>                   <NA>
#> 2                  <NA>                            <NA>                   <NA>
#> 3                  <NA>                            <NA>                   <NA>
#> 4                  <NA>                            <NA>                   <NA>
#> 5                  <NA>                            <NA>                   <NA>
#> 6                  <NA>                            <NA>                   <NA>
#>   location.fax_number               name
#> 1                <NA> Antitrust Division
#> 2                <NA> Antitrust Division
#> 3                <NA> Antitrust Division
#> 4                <NA> Antitrust Division
#> 5                <NA> Antitrust Division
#> 6                <NA> Antitrust Division